requirements.txt
requests
pyyaml
pandas
plotly-express
streamlit
streamlit-aggrid
st_btn_select

